Add README.md
The README.md provides some more context for repository visitors
and some information about how to use this. Currently, this
information is very fitted for our usage at EDYO, but I've tried
to keep it generalistic as possible.

This also includes renaming the `podcast.yaml` to
`podcast.yaml.sample` to avoid confusion.