Add tests for APPU 2
This adds `pytest` testing for the new `appu` subcommand, and the code
provided. Those tests are also included in the `test` pipeline to test
the subcommand within the CI pipeline. This uses a container to run
the tests, so it also validates building the APPU 2's container.

All the other steps in the CI pipeline, are selected to skip when
requesting to pull from a branch with `revamp` in the name. Ideally,
this should prevent issues between revamping changes and other
possible changes coming in.

The `PYTHONWARNINGS` environment variable in the Dockerfile is
provided to prevent weird output caused by [an issue with `pydub` and
regular expression support for Python
3.12](https://github.com/jiaaro/pydub/issues/801).